+++ src/rex.core/NEWS.rst
@@ -8,8 +8,9 @@
 1.1.0 (2013-XX-XX)
 ==================
 
+* Validators can now parse YAML documents.
 * ``RecordVal`` supports field names that coincide with Python keywords.
-* Added ``LatentRex`` class (Kyrylo Simonov).
+* Added ``LatentRex`` class.
 * Added NIH acknowledgement (Clark Evans).